+++ frontend/src/shared/api/index.ts
@@ -1,3 +1,4 @@
-export { default as AirportService } from './AirportService';
-export { airportHandlers } from './mock/airportHandler';
-export { flightHandlers } from './mock/flightHandler';
+export { default as AirportService } from './service/airport-service';
+export { airportHandlers } from './handler/airport-handler';
+export { flightHandlers } from './handler/flightHandler';
+export {airportMock} from './mock/airport-mock';
