+++ frontend/src/shared/api/mock/airport-mock.ts
@@ -1,7 +1,6 @@
-import type { AirportDto } from '../../types/api';
-import { http, HttpResponse } from 'msw';
+import type { AirportDto } from "../../types/api";
 
-export const airportsMock: AirportDto[] = [
+export const airportMock: AirportDto[] = [
   { id: 'svo', airport: 'Шереметьево', city: 'Москва' },
   { id: 'dme', airport: 'Домодедово', city: 'Москва' },
   { id: 'vko', airport: 'Внуково', city: 'Москва' },
@@ -16,21 +15,4 @@ export const airportsMock: AirportDto[] = [
   { id: 'vvo', airport: 'Кневичи', city: 'Владивосток' },
   { id: 'noz', airport: 'Спиченково', city: 'Новокузнецк' },
   { id: 'rof', airport: 'Платов', city: 'Ростов-на-Дону' },
-];
-
-export const airportHandlers = [
-  http.get('/api/airports', ({ request }) => {
-    const url = new URL(request.url);
-    const name = url.searchParams.get('name')?.toLowerCase().trim();
-
-    const filteredAirports = name
-      ? airportsMock.filter((item) => {
-          return (
-            item.airport.toLowerCase().includes(name) || item.city.toLowerCase().includes(name)
-          );
-        })
-      : airportsMock;
-
-    return HttpResponse.json(filteredAirports);
-  }),
 ];
